Time: 20 ms (15.68%), Space: 5.9 MB (73.52%) - LeetHub

+++ 69-sqrtx/69-sqrtx.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int mySqrt(int x) {
-        long long i = 1; long long val = 1;
+        long i = 1; long val = 1;
         while(val <= x){
             i++;
             val = i*i;
